requirements.txt
tensorflow-gpu==1.8.0
numpy
scipy
tqdm
opencv-python
keras
tensorboard
gitpython